refactor: Move tx-builder tests to sui_types and add doc
